# server/docs/swagger.yaml
definitions:
...
  nfc_relay_admin.ComplianceRuleRequest:
    properties:
      action:
        enum:
        - BLOCK
        - WARN
        - LOG
        type: string
      description:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      pattern:
        type: string
      riskLevel:
        enum:
        - LOW
        - MEDIUM
        - HIGH
        - CRITICAL
        type: string
    required:
    - action
    - name
    - riskLevel
    type: object
  nfc_relay_admin.ComplianceRuleResponse:
    properties:
      action:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      pattern:
        type: string
      riskLevel:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
  nfc_relay_admin.ComplianceRulesListResponse:
    properties:
      page:
        type: integer
      rules:
        items:
          $ref: '#/definitions/nfc_relay_admin.ComplianceRuleResponse'
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
...
  nfc_relay_admin.RuleTestRequest:
    properties:
      ruleId:
        type: string
      testContext:
        additionalProperties: true
        type: object
      testData:
        $ref: '#/definitions/security.APDUDataClass'
    required:
    - ruleId
    - testData
    type: object
  nfc_relay_admin.RuleTestResponse:
    properties:
      details:
        type: string
      matched:
        type: boolean
      result:
        $ref: '#/definitions/security.ComplianceResult'
      ruleId:
        type: string
      testTime:
        description: 微秒
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  security.APDUDataClass:
    properties:
      auditData:
        allOf:
        - $ref: '#/definitions/security.AuditableData'
        description: 审计数据 - 服务器可读，用于合规检查
      businessData:
        allOf:
        - $ref: '#/definitions/security.EncryptedBusinessData'
        description: 业务数据 - 服务器可解密，用于合规检查
      metadata:
        allOf:
        - $ref: '#/definitions/security.APDUMetadata'
        description: 元数据 - 明文，用于路由和监控
    type: object
  security.APDUMetadata:
    properties:
      checksumCrc:
        description: 数据完整性校验
        type: string
      clientId:
        type: string
      deviceInfo:
        description: 设备指纹
        type: string
      direction:
        description: '"command" 或 "response"'
        type: string
      sequenceNum:
        description: 序列号，防重放
        type: integer
      sessionId:
        type: string
      timestamp:
        type: string
      userId:
        type: string
    type: object
  security.AuditableData:
    properties:
      amount:
        description: 交易金额 (分为单位)
        type: integer
      applicationId:
        description: 应用ID (脱敏后)
        type: string
      commandClass:
        description: 命令类别 (SELECT, READ, WRITE等)
        type: string
      commandType:
        description: 具体命令类型
        type: string
      currency:
        description: 货币代码
        type: string
      merchantCategory:
        description: 商户类别
        type: string
      riskScore:
        description: 风险评分
        type: integer
      timestamp:
        description: 时间戳
        type: string
      transactionType:
        description: 交易类型
        type: string
    type: object
  security.ComplianceResult:
    properties:
      actions:
        description: 建议的处理动作
        items:
          type: string
        type: array
      compliant:
        type: boolean
      reason:
        type: string
      riskLevel:
        description: LOW, MEDIUM, HIGH, CRITICAL
        type: string
    type: object
  security.EncryptedBusinessData:
    properties:
      encryptedData:
        description: 所有业务数据都使用审计密钥加密，服务器可解密进行合规检查
        type: string
      encryptionInfo:
        allOf:
        - $ref: '#/definitions/security.EncryptionInfo'
        description: 加密元信息
    type: object
  security.EncryptionInfo:
    properties:
      algorithm:
        description: 加密算法
        type: string
      nonce:
        description: 随机数
        type: string
      tag:
        description: 认证标签
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.2
paths:
...
  /admin/nfc-relay/v1/compliance/rule-files/export:
    get:
      consumes:
      - application/json
      description: 导出当前的合规规则到文件
      parameters:
      - description: 导出格式
        enum:
        - json
        - yaml
        in: query
        name: format
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 导出规则文件
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/compliance/rule-files/import:
    post:
      consumes:
      - multipart/form-data
      description: 从文件导入合规规则
      parameters:
      - description: 规则文件
        in: formData
        name: file
        required: true
        type: file
      - description: 是否替换现有规则
        in: formData
        name: replace
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 导入规则文件
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/compliance/rules:
    get:
      consumes:
      - application/json
      description: 获取系统中的所有合规规则
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: size
        type: integer
      - description: 是否启用
        in: query
        name: enabled
        type: boolean
      - description: 风险级别
        in: query
        name: riskLevel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/nfc_relay_admin.ComplianceRulesListResponse'
              type: object
      summary: 获取所有合规规则
      tags:
      - NFC中继管理
    post:
      consumes:
      - application/json
      description: 创建新的合规规则
      parameters:
      - description: 合规规则信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/nfc_relay_admin.ComplianceRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建合规规则
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/compliance/rules/{rule_id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的合规规则
      parameters:
      - description: 规则ID
        in: path
        name: rule_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除合规规则
      tags:
      - NFC中继管理
    get:
      consumes:
      - application/json
      description: 根据ID获取特定的合规规则
      parameters:
      - description: 规则ID
        in: path
        name: rule_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/nfc_relay_admin.ComplianceRuleResponse'
              type: object
      summary: 获取单个合规规则
      tags:
      - NFC中继管理
    put:
      consumes:
      - application/json
      description: 更新现有的合规规则
      parameters:
      - description: 规则ID
        in: path
        name: rule_id
        required: true
        type: string
      - description: 合规规则信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/nfc_relay_admin.ComplianceRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新合规规则
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/compliance/rules/test:
    post:
      consumes:
      - application/json
      description: 使用测试数据验证合规规则的效果
      parameters:
      - description: 规则测试请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/nfc_relay_admin.RuleTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/nfc_relay_admin.RuleTestResponse'
              type: object
      summary: 测试合规规则
      tags:
      - NFC中继管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...